Refactor: move model common functions fmiFunctions.c to model_common.c
