+++ include/namespace.h
@@ -1,0 +1,12 @@
+#ifndef namespace_h
+#define namespace_h
+
+// C-code FMUs have functions names prefixed with MODEL_IDENTIFIER_.
+// Define DISABLE_PREFIX to build a binary FMU.
+#if !defined(DISABLE_PREFIX) && !defined(FMI3_FUNCTION_PREFIX)
+#define pasteA(a,b)          a ## b
+#define pasteB(a,b)          pasteA(a,b)
+#define FMI3_FUNCTION_PREFIX pasteB(MODEL_IDENTIFIER, _)
+#endif
+
+#endif /*namespace_h*/
